+++ crm/lib/integration/rest/configuration/entity/webform.php
@@ -2,8 +2,11 @@
 
 namespace Bitrix\Crm\Integration\Rest\Configuration\Entity;
 
+use Bitrix\Crm\Integration\Rest\Configuration\Helper;
+use Bitrix\Crm\Service\Container;
 use Bitrix\Rest;
 use Bitrix\Crm;
+use CCrmOwnerType;
 
 /**
  * Class WebForm
@@ -18,7 +21,6 @@ class WebForm
 	private $accessManifest = [
 		'crm_form',
 	];
-
 
 	/**
 	 * Get instance.
@@ -35,6 +37,35 @@ class WebForm
 		return self::$instance;
 	}
 
+	private function getDynamicTypeCustomSectionIdBySchemeValue(int $schemeValue): int
+	{
+		$result = 0;
+
+		$dynamicEntityTypeId = $this->getDynamicEntityTypeIdBySchemeValue($schemeValue);
+		if (
+			CCrmOwnerType::IsDefined($dynamicEntityTypeId)
+			&& CCrmOwnerType::isPossibleDynamicTypeId($dynamicEntityTypeId)
+		)
+		{
+			$type = Container::getInstance()->getTypeByEntityTypeId($dynamicEntityTypeId);
+			if ($type)
+			{
+				$customSectionId = $type->getCustomSectionId();
+				if ($customSectionId > 0)
+				{
+					$result = $customSectionId;
+				}
+			}
+		}
+
+		return $result;
+	}
+
+	private function getDynamicTypeCustomSectionIdByFormOptions(array $formOptions): int
+	{
+		return $this->getDynamicTypeCustomSectionIdBySchemeValue($this->getSchemeValue($formOptions));
+	}
+
 	/**
 	 * Export.
 	 *
@@ -48,7 +79,11 @@ class WebForm
 			return null;
 		}
 
-		$content = ['list' => []];
+		$helper = new Helper();
+		$content = [
+			'list' => [],
+			'dynamicTypesInfo' => $helper->exportCrmDynamicTypesInfo(),
+		];
 		$list = Crm\WebForm\Internals\FormTable::getDefaultTypeList([
 			'select' => ['ID'],
 			'filter' => [
@@ -58,8 +93,11 @@ class WebForm
 		]);
 		foreach ($list as $item)
 		{
-			$options = Crm\WebForm\Options::create($item['ID'])->getArray();
-			$content['list'][] = self::cleanFormOptions($options);
+			$options = self::cleanFormOptions(Crm\WebForm\Options::create($item['ID'])->getArray());
+			if ($this->getDynamicTypeCustomSectionIdByFormOptions($options) <= 0)
+			{
+				$content['list'][] = $options;
+			}
 		}
 
 		return [
@@ -85,14 +123,17 @@ class WebForm
 		if ($option['CLEAR_FULL'])
 		{
 			$list = Crm\WebForm\Internals\FormTable::getDefaultTypeList([
-				'select' => ['ID'],
+				'select' => ['ID', 'ENTITY_SCHEME'],
 				'filter' => [
 					'=IS_SYSTEM' => 'N',
 				],
 			]);
 			foreach ($list as $item)
 			{
-				Crm\WebForm\Form::delete($item['ID']);
+				if ($this->getDynamicTypeCustomSectionIdBySchemeValue((int)($item['ENTITY_SCHEME'] ?? 0)) <= 0)
+				{
+					Crm\WebForm\Form::delete($item['ID']);
+				}
 			}
 		}
 
@@ -127,6 +168,7 @@ class WebForm
 		foreach ($data['list'] as $options)
 		{
 			$options = self::cleanFormOptions($options);
+			$options = $this->prepareFormOptions($options, $import);
 			$options = Crm\WebForm\Options::createFromArray($options);
 			$options->getForm()->merge([
 				'ACTIVE' => 'Y',
@@ -138,7 +180,7 @@ class WebForm
 		return $result;
 	}
 
-	private static function cleanFormOptions(array $options)
+	private function cleanFormOptions(array $options)
 	{
 		$options['id'] = null;
 		$options['captcha'] = [];
@@ -158,5 +200,249 @@ class WebForm
 		);
 
 		return $options;
+	}
+
+	private function getSchemeValue(array $data): int
+	{
+		if (isset($data['document']['scheme']) && $data['document']['scheme'] > 0)
+		{
+			return (int)$data['document']['scheme'];
+		}
+
+		return 0;
+	}
+
+	private function getSchemeValueReminder(int $schemeValue)
+	{
+		return $schemeValue % 10;
+	}
+
+	private function getDynamicEntityTypeIdBySchemeValue(int $schemeValue): int
+	{
+		$result = CCrmOwnerType::Undefined;
+
+		if ($schemeValue > 10)
+		{
+			$result = (int)(($schemeValue - ($this->getSchemeValueReminder($schemeValue))) / 10);
+		}
+
+		return $result;
+	}
+
+	private function getNewDynamicEntityTypeIdByScheme(array $data, array $importData): int
+	{
+		$oldDynamicEntityTypeId = $this->getDynamicEntityTypeIdBySchemeValue($this->getSchemeValue($data));
+		$ratioKey = "DYNAMIC_$oldDynamicEntityTypeId";
+		if (isset($importData['RATIO']['CRM_DYNAMIC_TYPES'][$ratioKey]))
+		{
+			$newDynamicEntityTypeId = (int)$importData['RATIO']['CRM_DYNAMIC_TYPES'][$ratioKey];
+			if (CCrmOwnerType::isPossibleDynamicTypeId($newDynamicEntityTypeId))
+			{
+				return $newDynamicEntityTypeId;
+			}
+		}
+
+		return CCrmOwnerType::Undefined;
+	}
+
+	private function replaceDynamicScheme(array $data, array $importData): array
+	{
+		$newDynamicEntityTypeId = $this->getNewDynamicEntityTypeIdByScheme($data, $importData);
+		if ($newDynamicEntityTypeId > 0)
+		{
+			$schemeValue = $this->getSchemeValue($data);
+			if ($schemeValue > 10)
+			{
+				$reminder = $this->getSchemeValueReminder($schemeValue);
+				$data['document']['scheme'] = $newDynamicEntityTypeId * 10 + $reminder;
+			}
+		}
+
+		return $data;
+	}
+
+	private function replaceDynamicCategory(array $data, array $importData, int $oldDynamicEntityTypeId): array
+	{
+		if (
+			isset($data['document']['dynamic']['category'])
+			&& $data['document']['dynamic']['category'] > 0
+		)
+		{
+			$oldCategoryId = (int)$data['document']['dynamic']['category'];
+			$ratioKey = "DT{$oldDynamicEntityTypeId}_$oldCategoryId";
+			if (
+				isset($importData['RATIO']['CRM_STATUS'][$ratioKey])
+				&& $importData['RATIO']['CRM_STATUS'][$ratioKey] > 0
+			)
+			{
+				$newCategoryId = (int)$importData['RATIO']['CRM_STATUS'][$ratioKey];
+				$data['document']['dynamic']['category'] = $newCategoryId;
+			}
+		}
+
+		return $data;
+	}
+
+	private function prepareDynamicTypeReplacementLists(array $dynamicTypesInfo, array $ratioInfo): array
+	{
+		$result = [
+			'from' => [],
+			'to' => [],
+		];
+
+		foreach ($dynamicTypesInfo as $oldDynamicTypeId => $dynamicTypeInfo)
+		{
+			$oldDynamicTypeId = (int)$oldDynamicTypeId;
+			$oldDynamicEntityTypeId = (int)($dynamicTypeInfo['entityTypeId'] ?? 0);
+			if (
+				$oldDynamicEntityTypeId > 0
+				&& isset($ratioInfo['CRM_DYNAMIC_TYPES'])
+			)
+			{
+				$newDynamicTypeIdRatioKey = "DT$oldDynamicTypeId";
+				$newEntityTypeIdRatioKey = CCrmOwnerType::DynamicTypePrefixName . $oldDynamicEntityTypeId;
+				if (
+					isset($ratioInfo['CRM_DYNAMIC_TYPES'][$newDynamicTypeIdRatioKey])
+					&& $ratioInfo['CRM_DYNAMIC_TYPES'][$newDynamicTypeIdRatioKey] > 0
+					&& isset($ratioInfo['CRM_DYNAMIC_TYPES'][$newEntityTypeIdRatioKey])
+					&& $ratioInfo['CRM_DYNAMIC_TYPES'][$newEntityTypeIdRatioKey] > 0
+				)
+				{
+					$newDynamicTypeId = (int)$ratioInfo['CRM_DYNAMIC_TYPES'][$newDynamicTypeIdRatioKey];
+					$newDynamicEntityTypeId = (int)$ratioInfo['CRM_DYNAMIC_TYPES'][$newEntityTypeIdRatioKey];
+					
+					$oldDynamicEntityTypePrefix = CCrmOwnerType::DynamicTypePrefixName . $oldDynamicEntityTypeId;
+					$newDynamicEntityTypePrefix = CCrmOwnerType::DynamicTypePrefixName . $newDynamicEntityTypeId;
+
+					$matches = [];
+					if (
+						isset($dynamicTypeInfo['userFieldEntityId'])
+						&& is_string($dynamicTypeInfo['userFieldEntityId'])
+						&& isset($dynamicTypeInfo['userFieldNames'])
+						&& is_array($dynamicTypeInfo['userFieldNames'])
+						&& !empty($dynamicTypeInfo['userFieldNames'])
+						&& preg_match('/CRM_(\d+)/u', $dynamicTypeInfo['userFieldEntityId'], $matches)
+					)
+					{
+						$oldDynamicTypeId = $matches[1];
+						$oldUserFieldPrefix = "UF_CRM_{$oldDynamicTypeId}_";
+						$oldUserFieldPrefixLength = strlen($oldUserFieldPrefix);
+						$newUserFieldPrefix = "UF_CRM_{$newDynamicTypeId}_";
+						foreach($dynamicTypeInfo['userFieldNames'] as $oldUserFieldName)
+						{
+							if (substr($oldUserFieldName, 0, $oldUserFieldPrefixLength) === $oldUserFieldPrefix)
+							{
+								$newUserFieldName =
+									$newUserFieldPrefix . substr($oldUserFieldName, $oldUserFieldPrefixLength)
+								;
+
+								// /DYNAMIC_\d+_UF_CRM_\d+/
+								$result['from'][] = "{$oldDynamicEntityTypePrefix}_$oldUserFieldName";
+								$result['to'][] = "{$newDynamicEntityTypePrefix}_$newUserFieldName";
+
+								// /UF_CRM_\d+_/
+								$result['from'][] = $oldUserFieldName;
+								$result['to'][] = $newUserFieldName;
+							}
+						}
+					}
+
+					// /DYNAMIC_\d+/
+					$result['from'][] = $oldDynamicEntityTypePrefix;
+					$result['to'][] = $newDynamicEntityTypePrefix;
+				}
+			}
+		}
+
+		return $result;
+	}
+
+	private function getDynamicTypeReplacementLists(
+		array $dynamicTypesInfo,
+		array $ratioInfo,
+		bool $refresh = false
+	): array
+	{
+		static $replacementLists = null;
+
+		if ($replacementLists === null || $refresh)
+		{
+			$replacementLists = $this->prepareDynamicTypeReplacementLists($dynamicTypesInfo, $ratioInfo);
+		}
+
+		return $replacementLists;
+	}
+
+	private function changeDynamicTypeIdentifiers(
+		array|string $data,
+		array $replacementLists
+	): array|string
+	{
+		if (
+			isset($replacementLists['from'])
+			&& is_array($replacementLists['from'])
+			&& !empty($replacementLists['from'])
+			&& isset($replacementLists['to'])
+			&& is_array($replacementLists['to'])
+			&& !empty($replacementLists['to'])
+		)
+		{
+			if (is_string($data))
+			{
+				$replaceMarkers = [];
+				for ($i = 0; $i < count($replacementLists['from']); $i++)
+				{
+					$replaceMarkers[] = "_{<-rm[$i]->}_";
+				}
+				$data = str_replace($replacementLists['from'], $replaceMarkers, $data);
+				$data = str_replace($replaceMarkers, $replacementLists['to'], $data);
+			}
+			elseif (is_array($data))
+			{
+
+				foreach ($data as $key => $value)
+				{
+					$newKey = $this->changeDynamicTypeIdentifiers($key, $replacementLists);
+					if ($newKey !== $key)
+					{
+						unset($data[$key]);
+					}
+
+					if (is_string($value) || is_array($value))
+					{
+						$data[$newKey] = static::changeDynamicTypeIdentifiers($value, $replacementLists);
+					}
+					else
+					{
+						$data[$newKey] = $value;
+					}
+				}
+			}
+		}
+
+		return $data;
+	}
+
+	private function prepareFormOptions(
+		array|string $data,
+		array $importData,
+	): array|string
+	{
+
+		$oldDynamicEntityTypeId = $this->getDynamicEntityTypeIdBySchemeValue($this->getSchemeValue($data));
+		$data = $this->replaceDynamicScheme($data, $importData);
+		$data = $this->replaceDynamicCategory($data, $importData, $oldDynamicEntityTypeId);
+		$dynamicTypesInfo = $importData["CONTENT"]["DATA"]["dynamicTypesInfo"] ?? [];
+		if (is_array($dynamicTypesInfo) && !empty($dynamicTypesInfo))
+		{
+			$isSetRatio = (isset($importData['RATIO']) && is_array($importData['RATIO']));
+			$replacementLists = $this->getDynamicTypeReplacementLists(
+				$dynamicTypesInfo,
+				$isSetRatio ? $importData['RATIO'] : []
+			);
+			$data = $this->changeDynamicTypeIdentifiers($data, $replacementLists);
+		}
+
+		return $data;
 	}
 }
